Move iotest code into sharedtest
